Move QuantumClassifierWithDefaultRiemanianPipeline to pipelines module (#159)
* Create pipelines.py

* add pipeline module and move QuantumClassifierWithRiemannianPipeline to it

* [pre-commit.ci] auto fixes from pre-commit.com hooks

* flake8

* [pre-commit.ci] auto fixes from pre-commit.com hooks

* flake8

* fix dockerfile

* fix dockerfile

* fix import

* [pre-commit.ci] auto fixes from pre-commit.com hooks

* flake8

* [pre-commit.ci] auto fixes from pre-commit.com hooks

* typo

* fix import in example

---------
