+++ pyriemann_qiskit/__init__.py
@@ -1,4 +1,5 @@
 from ._version import __version__
 from . import classification
+from . import pipelines
 
-__all__ = ["__version__", "classification"]
+__all__ = ["__version__", "classification", "pipelines"]
